chore: refactor!: migrate `ClockService` to `com.lsadf.core.application.clock` package
- Relocated `ClockService`, `ClockServiceImpl`, and associated classes from `com.lsadf.core.infra.clock` to `com.lsadf.core.application.clock` for improved modular organization.
- Updated all imports across modules to reflect the new package structure.
- Annotated nullable fields and parameters in affected classes with `@Nullable` to enhance null safety.
- Added `package-info.java` with appropriate Apache License headers to the new package.
