+++ lsadf_api/src/test/java/com/lsadf/application/bdd/BddLoader.java
@@ -34,6 +34,7 @@ import com.lsadf.application.controller.user.UserController;
 import com.lsadf.application.controller.user.UserControllerImpl;
 import com.lsadf.config.LsadfConfiguration;
 import com.lsadf.core.application.cache.CacheManager;
+import com.lsadf.core.application.clock.ClockService;
 import com.lsadf.core.application.game.inventory.InventoryService;
 import com.lsadf.core.application.game.save.GameSaveService;
 import com.lsadf.core.application.game.save.characteristics.CharacteristicsService;
@@ -49,7 +50,6 @@ import com.lsadf.core.domain.game.save.metadata.GameMetadata;
 import com.lsadf.core.domain.game.save.stage.Stage;
 import com.lsadf.core.domain.user.User;
 import com.lsadf.core.domain.user.UserInfo;
-import com.lsadf.core.infra.clock.ClockService;
 import com.lsadf.core.infra.persistence.impl.game.inventory.AdditionalItemStatsRepository;
 import com.lsadf.core.infra.persistence.impl.game.inventory.ItemRepository;
 import com.lsadf.core.infra.persistence.impl.game.save.characteristics.CharacteristicsRepository;
