+++ lsadf_core/src/main/java/com/lsadf/core/application/clock/ClockService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.core.infra.clock;
+package com.lsadf.core.application.clock;
 
 import java.time.Clock;
 import java.time.Instant;
