+++ lsadf_core/src/main/java/com/lsadf/core/application/user/config/UserConfiguration.java
@@ -15,9 +15,9 @@
  */
 package com.lsadf.core.application.user.config;
 
+import com.lsadf.core.application.clock.ClockService;
 import com.lsadf.core.application.user.UserService;
 import com.lsadf.core.application.user.impl.UserServiceImpl;
-import com.lsadf.core.infra.clock.ClockService;
 import com.lsadf.core.infra.web.config.keycloak.properties.KeycloakProperties;
 import org.keycloak.admin.client.Keycloak;
 import org.springframework.context.annotation.Bean;
