+++ lsadf_core/src/main/java/com/lsadf/core/infra/logging/LoggingConfiguration.java
@@ -15,7 +15,7 @@
  */
 package com.lsadf.core.infra.logging;
 
-import com.lsadf.core.infra.clock.ClockService;
+import com.lsadf.core.application.clock.ClockService;
 import com.lsadf.core.infra.logging.interceptor.RequestLoggerInterceptor;
 import com.lsadf.core.infra.logging.properties.HttpLogProperties;
 import org.springframework.context.annotation.Bean;
